third_party/blink/web_tests/platform/generic/http/tests/websocket/url-attribute-expected.txt

.url attribute should work, previously supported .URL should not

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ws.url is "ws://127.0.0.1:8880/echo"
PASS ws.URL is undefined.
PASS successfullyParsed is true

TEST COMPLETE
